Fix path traversal tests with CC
CC changes the way inputs are represented so the unpack exception is no
longer thrown when checking if the inputs are empty (but no actual
unpacking happens there, so it is still safe). This allows it to proceed
to snapshot the output. Having '.' as an output is problematic on
Windows, because recursive snapshotting hits the locked file in
'.gradle'.

Using build output dir as a target is an easy workaround. FWIW, copying
anything into '.' fails at output snapshotting even without CC.